feat: unmanaged hostinterface

+++ controllers/daemon_watcher.go
@@ -246,9 +246,13 @@ func (w *DaemonWatcher) ProcessPodQueue() {
 			if errors.IsNotFound(err) {
 				// deleted, delete HostInterface
 				w.DaemonCacheHandler.SafeCache.UnsetCache(nodeName)
-				err := w.HostInterfaceHandler.DeleteHostInterface(nodeName)
-				if err != nil {
-					vars.DaemonLog.V(4).Info(fmt.Sprintf("Failed to delete HostInterface %s: %v", nodeName, err))
+				hif, err := w.HostInterfaceHandler.GetHostInterface(nodeName)
+				if err == nil && !vars.IsUnmanaged(hif.ObjectMeta) {
+					// deleted, delete HostInterface
+					err := w.HostInterfaceHandler.DeleteHostInterface(nodeName)
+					if err != nil {
+						vars.DaemonLog.V(4).Info(fmt.Sprintf("Failed to delete HostInterface %s: %v", nodeName, err))
+					}
 				}
 			}
 		}
